Eliminate one mark.clone() per character in Repetition
It doesn't matter what the value of self.marked is during the recursive
shift() call: we have a mutable reference to self, so we know that
nobody else can look at our data until after we return. So temporarily
write None there until we have the new result, and overwrite it then.

When I wrote this three years ago I apparently didn't know about
Option::take and punted on transferring ownership of the mark into the
recursive call to shift.