Convince myself that Sequence is correct
I couldn't decide whether this code was right, and comparing it to the
original "Play on Regular Expressions" paper didn't make me any less
confused.

So here's a giant pile of comments, lint annotations, and newtype
wrappers that prohibit all of the other ways of writing it that I
previously thought might have been correct.

I'm pretty sure this is right, now.