Rename all regex constructors, simplifying use
Overload |,&,!,+ operators for regex construction, and provide new
smart constructors empty(), is(), and many() where there aren't sensible
operators.

To get the operator implementations to type-check, I had to add newtype
wrappers and phantom data all over the place. But the exposed API is
nicer, so it was probably worth it.