Allow dynamic construction of grammars
This patch re-implements the Regex trait for boxed trait objects, which
I removed in d5ced3fa3ddbe13b3c37368a2b2ff35e90ffb371, because I have a
use for it now.

However, it was trickier this time, because cloning doesn't work in a
trait object. Cloning a grammar requires that the grammar's size is
known to the caller at compile-time, and trait objects are unsized.

So I've split the low-level clone implementations out of the Regex trait
into a new CloneRegex trait; the former supports trait objects and the
latter doesn't.

As a result, there's now an implementation of Box<Regex>, but not of
Box<CloneRegex>.