Allow lazily-constructed grammars
This allows building a large grammar if there are portions of the
grammar that aren't visited on some inputs.

In the paper "A Play on Regular Expressions", the authors observe that
lazily-evaluated grammars allow this kind of "regular expression"
matcher to match arbitrary context-free grammars as well. Since we
support intersection and complement, this implementation extends that
result to Boolean grammars.

That said, I'm not sure the implementation is actually lazy enough yet.
It should be partway there since I've implemented the "active" flag
approach from the paper, but at the moment, trying to check empty() will
force too much of the grammar to be evaluated. So that's a TODO...