Exhaustively test balanced_parentheses up to length 8
Generating random sequences of open parens and close parens very rarely
produces balanced parentheses, so quickcheck was almost exclusively
checking non-matching strings, with the occasional match on the empty
string.

As a mathematical aside, a friend tells me:

> For a string of length 2n the number of balanced strings is described
> by the nth Catalan number, and grows ~ n^(-3/2) as a fraction of the
> total possible strings.

On the other hand, generating and testing all strings of length <= 8 on
an alphabet of "()x" takes very little time and covers a range of
strings that should match while still producing plenty that don't.