Bugfix: Grammars are active if any child is
I got this wrong for And grammars. If at some point you start only
shifting in zero forever, and one of the children is inactive, then it's
true that an And grammar will never shift out a non-zero. But just
because you shift in a zero now doesn't mean all future incoming marks
will be zero, and in case you shift in a non-zero in the future, the
active child still needs to shift its marks around.